refactor: move events to a new file
