+++ src-tauri/src/window/mod.rs
@@ -1,1 +1,2 @@
+pub mod events;
 pub mod window_management;
